快速排序

+++ J20230911/src/practice/Demo2.java
@@ -53,7 +53,7 @@ public class Demo2 {
         System.out.println();
     }
 
-    public int[] fibonacci(int n) {
+    public long[] fibonacci(int n) {
         long[] fibArray = new long[n + 1];
         fibArray[0] = 0;
         fibArray[1] = 1;
